# impl/gppractice/go-gin-server/go/api_gp_practice.go
/*
 * GP Practice API
 *
 * ## API for GP Practice lookups
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package gprecords

import (
	"go-gin-server/data"
	"net/http"
	"log"	
	"github.com/gin-gonic/gin"
)

// GetAllGPRecords - Get All GP Practices
func GetAllGPRecords(c *gin.Context) {
	gpc := data.GetGPRecords()
	c.JSON(http.StatusOK, data.GPRecordsToJSON(gpc))
}

// GetGPRecord - Get GP Practices
func GetGPRecord(c *gin.Context) {
	gmcCode := c.Query("gmcCode")
	log.Println(gmcCode)
	byGmc,_ := data.GetGPRecordsByGMC(gmcCode)
	c.JSON(http.StatusOK, data.GPRecordsToJSON(byGmc))
}